+++ core/src/main/java/eu/kennytv/maintenance/core/Settings.java
@@ -48,7 +48,7 @@ import org.jetbrains.annotations.Nullable;
 public class Settings implements eu.kennytv.maintenance.api.Settings {
     public static final String NEW_LINE_REPLACEMENT = "<br>";
     private static final int CONFIG_VERSION = 11;
-    private static final int LANGUAGE_VERSION = 3;
+    private static final int LANGUAGE_VERSION = 4;
     protected final MaintenancePlugin plugin;
     private final Map<UUID, String> whitelistedPlayers = new HashMap<>();
     private final String[] unsupportedFields;